refactor(painting): make searchBar component to be related searching painting
